requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
Werkzeug
psycopg2-binary
flask-cors